Test/fixes and improvements rebase

+++ src-java/testing/functional-tests/src/test/groovy/org/openkilda/functionaltests/spec/flows/BandwidthV2Spec.groovy
@@ -5,19 +5,24 @@ import static org.openkilda.functionaltests.extension.tags.Tag.SMOKE
 import static org.openkilda.testing.Constants.WAIT_OFFSET
 
 import org.openkilda.functionaltests.HealthCheckSpecification
+import org.openkilda.functionaltests.extension.failfast.Tidy
 import org.openkilda.functionaltests.extension.tags.Tags
 import org.openkilda.functionaltests.helpers.PathHelper
 import org.openkilda.functionaltests.helpers.Wrappers
+import org.openkilda.messaging.error.MessageError
 import org.openkilda.messaging.info.event.IslInfoData
 import org.openkilda.messaging.info.event.PathNode
 import org.openkilda.messaging.payload.flow.FlowState
 
+import org.springframework.http.HttpStatus
 import org.springframework.web.client.HttpClientErrorException
+import spock.lang.Ignore
 import spock.lang.Narrative
 
 @Narrative("Verify that ISL's bandwidth behaves consistently and does not allow any oversubscribtions etc.")
 class BandwidthV2Spec extends HealthCheckSpecification {
 
+    @Tidy
     @Tags(SMOKE)
     def "Available bandwidth on ISLs changes respectively when creating/updating/deleting a flow"() {
         given: "Two active not neighboring switches"
@@ -57,13 +62,17 @@ class BandwidthV2Spec extends HealthCheckSpecification {
                 maximumBandwidth - maximumBandwidthUpdated)
 
         when: "Delete the flow"
-        flowHelperV2.deleteFlow(flow.flowId)
+        def deleteResponse = flowHelperV2.deleteFlow(flow.flowId)
 
         then: "Available bandwidth on ISLs is changed to the initial value before flow creation"
         def linksAfterFlowDelete = northbound.getAllLinks()
         checkBandwidth(flowPathAfterUpdate, linksBeforeFlowCreate, linksAfterFlowDelete)
+
+        cleanup:
+        !deleteResponse && flowHelperV2.deleteFlow(flow.flowId)
     }
 
+    @Tidy
     def "Longer path is chosen in case of not enough available bandwidth on a shorter path"() {
         given: "Two active switches with two possible flow paths at least"
         def switchPair = topologyHelper.getAllNeighboringSwitchPairs().find { it.paths.size() > 1 } ?:
@@ -98,10 +107,13 @@ class BandwidthV2Spec extends HealthCheckSpecification {
         def flow2Path = PathHelper.convert(northbound.getFlowPath(flow2.flowId))
         pathHelper.getCost(flow2Path) > pathHelper.getCost(flow1Path)
 
-        and: "Delete created flows"
-        [flow1.flowId, flow2.flowId].each { flowHelperV2.deleteFlow(it) }
+        cleanup: "Delete created flows"
+        [flow1?.flowId, flow2?.flowId].each {
+            it && flowHelperV2.deleteFlow(it)
+        }
     }
 
+    @Tidy
     def "Unable to exceed bandwidth limit on ISL when creating a flow"() {
         given: "Two active switches"
         def switchPair = topologyHelper.getNeighboringSwitchPair()
@@ -121,8 +133,12 @@ class BandwidthV2Spec extends HealthCheckSpecification {
         then: "The flow is not created because flow path should not be found"
         def exc = thrown(HttpClientErrorException)
         exc.rawStatusCode == 404
+
+        cleanup:
+        !exc && flowHelperV2.deleteFlow(flow.flowId)
     }
 
+    @Tidy
     def "Unable to exceed bandwidth limit on ISL when updating a flow"() {
         given: "Two active switches"
         def switchPair = topologyHelper.getNeighboringSwitchPair()
@@ -150,10 +166,11 @@ class BandwidthV2Spec extends HealthCheckSpecification {
         def e = thrown(HttpClientErrorException)
         e.rawStatusCode == 404
 
-        and: "Delete the flow"
+        cleanup: "Delete the flow"
         flowHelperV2.deleteFlow(flow.flowId)
     }
 
+    @Tidy
     def "Able to exceed bandwidth limit on ISL when creating/updating a flow with ignore_bandwidth=true"() {
         given: "Two active switches"
         def switchPair = topologyHelper.getNeighboringSwitchPair()
@@ -190,10 +207,11 @@ class BandwidthV2Spec extends HealthCheckSpecification {
         flowPathAfterUpdate == flowPath
         checkBandwidth(flowPathAfterUpdate, linksBeforeFlowCreate, linksAfterFlowUpdate)
 
-        and: "Delete the flow"
+        cleanup: "Delete the flow"
         flowHelperV2.deleteFlow(flow.flowId)
     }
 
+    @Tidy
     def "Able to update bandwidth to maximum link speed without using alternate links"() {
         given: "Two active neighboring switches"
         def switchPair = topologyHelper.getNeighboringSwitchPair()
@@ -231,8 +249,54 @@ class BandwidthV2Spec extends HealthCheckSpecification {
         and: "The same path is used by updated flow"
         PathHelper.convert(northbound.getFlowPath(flow.flowId)) == flowPath
 
-        and: "Delete the flow"
+        cleanup: "Delete the flow"
         flowHelperV2.deleteFlow(flow.flowId)
+    }
+
+    @Ignore("https://github.com/telstra/open-kilda/issues/1150")
+    @Tidy
+    def "System doesn't allow to exceed bandwidth limit on ISL while updating a flow with ignore_bandwidth=false"() {
+        given: "Two active switches"
+        def switchPair = topologyHelper.getNeighboringSwitchPair()
+
+        when: "Create a flow with a bandwidth that exceeds available bandwidth on ISL (ignore_bandwidth=true)"
+        def linksBeforeFlowCreate = northbound.getAllLinks()
+        def flow = flowHelperV2.randomFlow(switchPair)
+        long maxBandwidth = northbound.getAllLinks()*.availableBandwidth.max()
+        flow.maximumBandwidth = maxBandwidth + 1
+        flow.ignoreBandwidth = true
+        flowHelperV2.addFlow(flow)
+        assert northbound.getFlow(flow.flowId).maximumBandwidth == flow.maximumBandwidth
+
+        then: "Available bandwidth on ISLs is not changed in accordance with flow maximum bandwidth"
+        def linksAfterFlowCreate = northbound.getAllLinks()
+        def flowPath = PathHelper.convert(northbound.getFlowPath(flow.flowId))
+        checkBandwidth(flowPath, linksBeforeFlowCreate, linksAfterFlowCreate)
+
+        when: "Update the flow (ignore_bandwidth = false)"
+        northboundV2.updateFlow(flow.flowId, flow.tap { it.ignoreBandwidth = false })
+
+        then: "Human readable error is returned"
+        def exc = thrown(HttpClientErrorException)
+        exc.statusCode == HttpStatus.NOT_FOUND
+        def errorDetails = exc.responseBodyAsString.to(MessageError)
+        errorDetails.errorMessage == "Could not update flow"
+        errorDetails.errorDescription == "Not enough bandwidth or no path found. " +
+                "Failed to find path with requested bandwidth=${flow.maximumBandwidth}: " +
+                "Switch ${flow.source.switchId} doesn't have links with enough bandwidth"
+
+        and: "The flow is not updated and has 'Up' status"
+        Wrappers.wait(WAIT_OFFSET) { assert northbound.getFlowStatus(flow.flowId).status == FlowState.UP }
+        northbound.getFlow(flow.flowId).ignoreBandwidth
+
+        and: "Available bandwidth on ISLs is not changed"
+        def linksAfterFlowUpdate = northbound.getAllLinks()
+        def flowPathAfterUpdate = PathHelper.convert(northbound.getFlowPath(flow.flowId))
+        flowPathAfterUpdate == flowPath
+        checkBandwidth(flowPathAfterUpdate, linksBeforeFlowCreate, linksAfterFlowUpdate)
+
+        cleanup: "Delete the flow"
+        flow && flowHelperV2.deleteFlow(flow.flowId)
     }
 
     def cleanup() {
